+++ src/components/select/select.styles.ts
@@ -176,6 +176,10 @@ export default css`
     margin-inline-end: var(--sl-input-spacing-small);
   }
 
+  .select--small .select__suffix::slotted(*) {
+    margin-inline-start: var(--sl-input-spacing-small);
+  }
+
   .select--small.select--multiple:not(.select--placeholder-visible) .select__combobox {
     padding-block: 2px;
     padding-inline-start: 0;
@@ -199,6 +203,10 @@ export default css`
 
   .select--medium .select__prefix::slotted(*) {
     margin-inline-end: var(--sl-input-spacing-medium);
+  }
+
+  .select--medium .select__suffix::slotted(*) {
+    margin-inline-start: var(--sl-input-spacing-medium);
   }
 
   .select--medium.select--multiple:not(.select--placeholder-visible) .select__combobox {
@@ -226,6 +234,10 @@ export default css`
     margin-inline-end: var(--sl-input-spacing-large);
   }
 
+  .select--large .select__suffix::slotted(*) {
+    margin-inline-start: var(--sl-input-spacing-large);
+  }
+
   .select--large.select--multiple:not(.select--placeholder-visible) .select__combobox {
     padding-inline-start: 0;
     padding-block: 4px;
@@ -249,14 +261,7 @@ export default css`
   }
 
   /* Prefix */
-  .select__prefix {
-    flex: 0;
-    display: inline-flex;
-    align-items: center;
-    color: var(--sl-input-placeholder-color);
-  }
-
-  /* Suffix */
+  .select__prefix,
   .select__suffix {
     flex: 0;
     display: inline-flex;